requirements.txt
# LinguaBot — Python dependencies
python-telegram-bot[job-queue]==21.3
aiohttp>=3.9,<4
aiosqlite>=0.20,<1
python-dotenv>=1.0,<2
